+++ sdk/core/azure-core/src/http/retry_policy.cpp
@@ -137,6 +137,7 @@ std::unique_ptr<RawResponse> Azure::Core::Http::RetryPolicy::Send(
     request.StartTry();
     // creates a copy of original query parameters from request
     auto originalQueryParameters = request.GetUrl().GetQueryParameters();
+
     try
     {
       auto response = nextHttpPolicy.Send(ctx, request);
@@ -150,17 +151,12 @@ std::unique_ptr<RawResponse> Azure::Core::Http::RetryPolicy::Send(
         return response;
       }
     }
-    catch (TransportException const&)
+    catch (const TransportException&)
     {
       if (!ShouldRetryOnTransportFailure(m_retryOptions, attempt, retryAfter))
       {
         throw;
       }
-    }
-
-    if (auto bodyStream = request.GetBodyStream())
-    {
-      bodyStream->Rewind();
     }
 
     if (shouldLog)