+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 12.0.0-beta.9 (Unreleased)
 
+### New Features
+
+- Added support for customized application ID.
+
+### Breaking Changes
+
+- DataLake client constructors won't automatically convert blob url to dfs url anymore.
 
 ## 12.0.0-beta.8 (2021-02-12)